+++ docs/reference/content/changelog.md
@@ -11,6 +11,7 @@ title = "Changelog"
 Changes between released versions
 
 ### 1.2.0
+  * Added support for library authors to extend the handshake metadata. [SCALA-252](https://jira.mongodb.org/browse/SCALA-252)
   * Added support for the new Aggregation stages in 3.4 [SCALA-258](https://jira.mongodb.org/browse/SCALA-258)
   * Added Collation support [SCALA-249](https://jira.mongodb.org/browse/SCALA-249)
   * Added support for BsonDecimal128 [SCALA-241](https://jira.mongodb.org/browse/SCALA-241)